refactor(ui): refactor Tooltip
